+++ lib/core.🫧
@@ -5,6 +5,10 @@
 # (mufn puts [& msg]
 #   (eval (push msg °console.log)))
 
+# (mufn puts [& msg]
+#   (console.log msg))
+# (mufn puts [& msg] msg)
+
 (muf 🍎 (fn [a b] (console.log a b)))
 
 (muf peek (fn [list] (send list :peek)))
